# spec/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ory/keto/opl/v1alpha1/syntax_service.proto",
    "version": "version not set"
  },
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/admin/relation-tuples": {
      "delete": {
        "summary": "Deletes relationships based on relation query",
        "operationId": "deleteRelationships",
        "responses": {
          "204": {
            "description": "An empty response.",
            "schema": {}
          }
        },
        "parameters": [
          {
            "name": "namespace",
            "description": "The namespace this relation tuple lives in.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "object",
            "description": "The object related by this tuple.\nIt is an object in the namespace of the tuple.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "relation",
            "description": "The relation between an Object and a Subject.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "subject_id",
            "description": "A concrete id of the subject.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "subject_set.namespace",
            "description": "The namespace of the object and relation\nreferenced in this subject set.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "subject_set.object",
            "description": "The object related by this subject set.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "subject_set.relation",
            "description": "The relation between the object and the subjects.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": ["relationship"],
        "consumes": ["application/x-www-form-urlencoded"]
      },
      "put": {
        "summary": "Creates a relationship",
        "operationId": "createRelationship",
        "responses": {
          "201": {
            "description": "The created relationship.",
            "schema": {
              "$ref": "#/definitions/relationship"
            }
          }
        },
        "parameters": [
          {
            "name": "relationTuple",
            "description": "The relationship to create.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/createRelationshipBody"
            }
          }
        ],
        "tags": ["relationship"]
      },
      "patch": {
        "summary": "Writes one or more relationships in a single transaction.",
        "operationId": "patchRelationships",
        "responses": {
          "204": {
            "description": "An empty response.",
            "schema": {},
            "examples": {
              "application/json": {}
            }
          }
        },
        "parameters": [
          {
            "name": "relationTupleDeltas",
            "description": "The write delta for the relationships operated in one single transaction.\nEither all actions succeed or no change takes effect on error.",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/relationshipDelta"
              }
            }
          }
        ],
        "tags": ["relationship"]
      }
    },
...
    "/relation-tuples": {
      "get": {
        "summary": "Lists ACL relationships.",
        "operationId": "getRelationships",
        "responses": {
          "200": {
            "description": "The list of relationships.",
            "schema": {
              "$ref": "#/definitions/relationships"
            }
          }
        },
        "parameters": [
          {
            "name": "page_size",
            "description": "Optional. The maximum number of\nRelationTuples to return in the response.\n\nDefault: 100",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "page_token",
            "description": "Optional. An opaque pagination token returned from\na previous call to `ListRelationTuples` that\nindicates where the page should start at.\n\nAn empty token denotes the first page. All successive\npages require the token from the previous page.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "namespace",
            "description": "The namespace",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "object",
            "description": "The related object in this check.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "relation",
            "description": "The relation between the Object and the Subject.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "subject_id",
            "description": "A concrete id of the subject.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "subject_set.namespace",
            "description": "The namespace of the object and relation\nreferenced in this subject set.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "subject_set.object",
            "description": "The object related by this subject set.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "subject_set.relation",
            "description": "The relation between the object and the subjects.",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": ["relationship"],
        "consumes": ["application/x-www-form-urlencoded"]
      }
    },
...
  },
  "definitions": {
...
    "createRelationshipBody": {
      "type": "object",
      "properties": {
        "namespace": {
          "type": "string",
          "description": "The namespace this relation tuple lives in."
        },
        "object": {
          "type": "string",
          "description": "The object related by this tuple.\nIt is an object in the namespace of the tuple."
        },
        "relation": {
          "type": "string",
          "description": "The relation between an Object and a Subject."
        },
        "subject_id": {
          "type": "string",
          "description": "A concrete id of the subject."
        },
        "subject_set": {
          "$ref": "#/definitions/subjectSet",
          "description": "A subject set that expands to more Subjects.\nMore information are available under [concepts](../concepts/subjects.mdx)."
        }
      }
    },
...
    "relationships": {
      "type": "object",
      "properties": {
        "relation_tuples": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/relationship"
          },
          "description": "The relationships matching the list request."
        },
        "next_page_token": {
          "type": "string",
          "description": "The token required to get the next page.\nIf this is the last page, the token will be the empty string."
        }
      },
      "description": "The response of a ReadService.ListRelationTuples RPC."
    },
...
    "relationship": {
      "type": "object",
      "properties": {
        "namespace": {
          "type": "string",
          "description": "The namespace this relation tuple lives in."
        },
        "object": {
          "type": "string",
          "description": "The object related by this tuple.\nIt is an object in the namespace of the tuple."
        },
        "relation": {
          "type": "string",
          "description": "The relation between an Object and a Subject."
        },
        "subject_id": {
          "type": "string",
          "description": "A concrete id of the subject."
        },
        "subject_set": {
          "$ref": "#/definitions/subjectSet",
          "description": "A subject set that expands to more Subjects.\nMore information are available under [concepts](../concepts/subjects.mdx)."
        }
      },
      "description": "RelationTuple defines a relation between an Object and a Subject.",
      "required": ["namespace", "object", "relation"]
    },
    "relationshipDelta": {
      "type": "object",
      "properties": {
        "action": {
          "$ref": "#/definitions/relationshipDelta.Action",
          "description": "The action to do on the RelationTuple."
        },
        "relation_tuple": {
          "$ref": "#/definitions/relationship",
          "description": "The target RelationTuple."
        }
      },
      "description": "Write-delta for a TransactRelationTuplesRequest."
    },
    "relationshipDelta.Action": {
      "type": "string",
      "enum": ["ACTION_UNSPECIFIED", "insert", "delete"],
      "default": "ACTION_UNSPECIFIED",
      "description": " - ACTION_UNSPECIFIED: Unspecified.\nThe `TransactRelationTuples` RPC ignores this\nRelationTupleDelta if an action was unspecified.\n - insert: Insertion of a new RelationTuple.\nIt is ignored if already existing.\n - delete: Deletion of the RelationTuple.\nIt is ignored if it does not exist."
    },
...
    "subjectSet": {
      "type": "object",
      "properties": {
        "namespace": {
          "type": "string",
          "description": "The namespace of the object and relation\nreferenced in this subject set."
        },
        "object": {
          "type": "string",
          "description": "The object related by this subject set."
        },
        "relation": {
          "type": "string",
          "description": "The relation between the object and the subjects."
        }
      },
      "description": "SubjectSet refers to all subjects who have\nthe same `relation` on an `object`.",
      "required": ["namespace", "object", "relation"]
    },
...
  }
}